+++ spring-cloud-tencent-examples/polaris-circuitbreaker-example/polaris-circuitbreaker-callee-service2/src/main/resources/bootstrap.yml
@@ -1,0 +1,13 @@
+server:
+  port: 48082
+spring:
+  application:
+    name: polaris-circuitbreaker-callee-service
+  cloud:
+    polaris:
+      address: grpc://127.0.0.1:8091
+      namespace: default
+      enabled: true
+      stat:
+        enabled: true
+        port: 28083
