--- spring-cloud-tencent-examples/polaris-circuitbreaker-example/polaris-circuitbreaker-example-a/src/main/java/com/tencent/cloud/polaris/circuitbreaker/example/ProviderB.java
@@ -1,40 +1,0 @@
-/*
- * Tencent is pleased to support the open source community by making Spring Cloud Tencent available.
- *
- * Copyright (C) 2019 THL A29 Limited, a Tencent company. All rights reserved.
- *
- * Licensed under the BSD 3-Clause License (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * https://opensource.org/licenses/BSD-3-Clause
- *
- * Unless required by applicable law or agreed to in writing, software distributed
- * under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
- * CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
-
-package com.tencent.cloud.polaris.circuitbreaker.example;
-
-import org.springframework.cloud.openfeign.FeignClient;
-import org.springframework.context.annotation.Primary;
-import org.springframework.web.bind.annotation.GetMapping;
-
-/**
- * Circuit breaker example callee provider.
- *
- * @author Haotian Zhang
- */
-@Primary
-@FeignClient(name = ProviderBFallbackConstant.SERVICE_NAME, fallback = ProviderBFallback.class)
-public interface ProviderB {
-
-	/**
-	 * Get info of service B.
-	 *
-	 * @return info of service B
-	 */
-	@GetMapping(ProviderBFallbackConstant.API)
-	String info();
-}
