--- spring-cloud-tencent-examples/polaris-circuitbreaker-example/polaris-circuitbreaker-example-a/src/main/resources/bootstrap.yml
@@ -1,43 +1,0 @@
-server:
-  port: 48080
-spring:
-  application:
-    name: polaris-circuitbreaker-example-a
-  cloud:
-    polaris:
-      address: grpc://127.0.0.1:8091
-      namespace: default
-      enabled: true
-      circuitbreaker:
-        enabled: true
-      stat:
-        enabled: true
-        port: 28081
-    loadbalancer:
-      configurations: polaris
-    tencent:
-      rpc-enhancement:
-        enabled: true
-        reporter:
-          ignore-internal-server-error: true
-          series: server_error
-          statuses: gateway_timeout, bad_gateway, service_unavailable
-
-feign:
-  circuitbreaker:
-    enabled: true
-  compression:
-    request:
-      enabled: false
-      mime-types: text/xml,application/xml,application/json
-      min-request-size: 2048
-    response:
-      enabled: false
-
-serivceB:
-  url: http://localhost:48081
-logging:
-  level:
-    root: info
-    com.tencent.cloud: debug
-
