--- spring-cloud-tencent-examples/polaris-circuitbreaker-example/polaris-circuitbreaker-example-b2/src/main/resources/bootstrap.yml
@@ -1,13 +1,0 @@
-server:
-  port: 48082
-spring:
-  application:
-    name: polaris-circuitbreaker-example-b
-  cloud:
-    polaris:
-      address: grpc://127.0.0.1:8091
-      namespace: default
-      enabled: true
-      stat:
-        enabled: true
-        port: 28083
