+++ spring-cloud-tencent-examples/polaris-circuitbreaker-example/polaris-circuitbreaker-gateway-example/src/main/resources/bootstrap.yml
@@ -1,0 +1,59 @@
+server:
+  session-timeout: 1800
+  port: 48080
+spring:
+  application:
+    name: GatewayScgService
+  cloud:
+    tencent:
+      plugin:
+        scg:
+          staining:
+            enabled: true
+            rule-staining:
+              enabled: true
+        router:
+          feature-env:
+            enabled: true
+    polaris:
+      address: grpc://127.0.0.1:8091
+      namespace: default
+      enabled: true
+    gateway:
+      discovery:
+        locator:
+          enabled: true
+          'predicates[0]':
+            name: Path
+            args:
+              patterns: '''/'' + serviceId + ''/**'''
+          'filters[0]':
+            name: RewritePath
+            args:
+              regexp: '''/'' + serviceId + ''/(?<remaining>.*)'''
+              replacement: '''/$\{remaining}'''
+          'filters[1]':
+            name: CircuitBreaker
+            args:
+              name: 'serviceId'
+              statusCodes: 502
+              fallbackUri: '''forward:/polaris-fallback'''
+#      routes:
+#        - id: polaris-circuitbreaker-callee-service
+#          uri: lb://polaris-circuitbreaker-callee-service
+#          predicates:
+#            - Path=/polaris-circuitbreaker-callee-service/**
+#          filters:
+#            - StripPrefix=1
+#            - name: CircuitBreaker
+#              args:
+#                name: polaris-circuitbreaker-callee-service
+#                statusCodes: 502
+#                fallbackUri: forward:/polaris-fallback
+logging:
+  level:
+    root: info
+    com.tencent.polaris.discovery.client.flow.RegisterFlow: off
+    com.tencent.polaris.plugins.registry.memory.CacheObject: off
+    com.tencent.cloud.polaris.circuitbreaker: debug
+
