+++ spring-cloud-tencent-examples/polaris-circuitbreaker-example/polaris-circuitbreaker-webclient-example/src/main/resources/bootstrap.yml
@@ -1,0 +1,30 @@
+server:
+  port: 48080
+spring:
+  application:
+    name: polaris-circuitbreaker-webclient-example
+  cloud:
+    polaris:
+      address: grpc://127.0.0.1:8091
+      namespace: default
+      enabled: true
+      loadbalancer:
+        enabled: true
+      circuitbreaker:
+        enabled: true
+#      stat:
+#        enabled: true
+#        port: 28081
+#    tencent:
+#      rpc-enhancement:
+#        enabled: true
+#        reporter:
+#          ignore-internal-server-error: true
+#          series: server_error
+#          statuses: gateway_timeout, bad_gateway, service_unavailable
+
+logging:
+  level:
+    root: info
+    com.tencent.cloud: debug
+
